refactor: extract ios lifecycle helpers
